==> 03_Setup_Run_Mortality_Prediction_Model.ipynb <==
<div class="alert alert-block alert-info">
    <b>Import Packages</b>
</div>


import pandas as  pd
import os
import sys
import numpy as np
import dtale
import json
from tqdm import tqdm

<div class="alert alert-block alert-info">
    <b>Import the local pyhospice module so it can be accessed like any other module</b>
</div>


sys.path.append('/pyhospice')

<div class="alert alert-block alert-info">
    <b>Jupyter Plumbing</b>
</div>


# %load C:\users\derek\.jupyter\startup.py
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

from IPython.core.display import display, HTML
display(HTML("<style>.container { width:100% !important; }</style>"))

#pd.set_option('display.max_colwidth', 100)
pd.set_option('display.max_colwidth', None)
pd.set_option("display.max_rows", None)
pd.options.display.max_columns = 500
pd.options.display.max_seq_items = 500

<div class="alert alert-block alert-info">
    <b>View .csv files in the _csvStore working directory</b>
</div>


csvStore = os.path.join(os.getcwd(),'__csvStore' )
csvlist = []
for file in enumerate(os.listdir(csvStore)):
    if file[1].endswith(".csv"):
        csvlist.append(os.path.join(str(csvStore), file[1]))
        print(os.path.join(str(csvStore), file[1]))

# enumerate csvlist for easier dynamic import of different versions
list(enumerate(csvlist))

<div class="alert alert-block alert-info">
    <b>Create df by import .csv files created in previous step</b>
</div>


df_patients = pd.read_csv(csvlist[12])#.iloc[:1000,:]
df_episodes = pd.read_csv(csvlist[1])#.iloc[:150,:]
df_visits = pd.read_csv(csvlist[0])#.iloc[:150,:]
df_OASIS = pd.read_csv(csvlist[3])#.iloc[:150,:]
df_event_mapping = pd.read_csv(csvlist[9])

<hr style="height:2px;border-width:0;color:black;background-color:black">

<div class="alert alert-block alert-info">
    <b>This allows us to filter our data to only a handful of patients for testing purposes</b>
</div>


dead = df_patients[df_patients.mortality == 1].pa_id
f"There are {len(dead)} patients"

df_OASIS_dead = df_OASIS[df_OASIS.pa_id.isin(dead)]
df_OASIS_dead.dtypes

<div class="alert alert-block alert-info">
    <b>Creates A Dictionary with {pa_id: row_numbers/num of rows} to see which patients have the most rows (most number of visits/data)</b>
</div>


grpz = df_OASIS_dead.groupby('pa_id').groups
#df_OASIS.groupby('pa_id').groups[36796]
#df_OASIS.groupby('pa_id').get_group(tup)

<div class="alert alert-block alert-info">
    <b>Find the Patient w the most Rows which are good candiates for testing</b>
</div>


def patient_most_rows(grpz):
    maxcount = max(len(v) for v in grpz.values())
    print(f'The max rownum is {maxcount}, and the pa_id is:')
    return [k for k, v in grpz.items() if len(v) == maxcount]


patient_most_rows(grpz)

length_dict = {k: len(v) for k, v in grpz.items()}
[k for k, v in length_dict.items() if v == max(length_dict.values())]
#length_dict

import heapq
dead100 = [k for k, v in length_dict.items() if v in heapq.nlargest(225, length_dict.values())]
len(dead100)

<div class="alert alert-block alert-info">
    <b>We will take all the 'Alive' patients since there are only 225</b>
</div>

alive = df_patients[df_patients.mortality==0].pa_id
f"There are, {len(alive)}, alive patients"

<div class="alert alert-block alert-info">
    <b>This is our testing patient subset</b>
</div>

pa_subset = alive.tolist() + dead100
f"For testing purposes we will use {len(set(pa_subset))} patient subset"

<div class="alert alert-block alert-info">
    <b>Final Filter for the main df: we can hard code specific patients if desired</b>
</div>

# Use this to hardcode patients you want to keep for testing
#grpidx = (17191,44654,25607,18808,16862,17018,24556,16274,26475,27968,36781,17747,37870,18011,30512,51094,38501,36509,17716,42124,16950,353347,326385,319247,318537,307867,283257,268061,347156,325432,348164,348170,350202,335984,359871,364367,320405,306152,322487,322263,308977,280963)

grpidx = pa_subset

<div class="alert alert-block alert-info">
    <b>Repopulate the original df with test data</b>
</div>


df_patients = df_patients[df_patients.pa_id.isin(grpidx)]
df_episodes = df_episodes[df_episodes.pa_id.isin(grpidx)]
df_visits = df_visits[df_visits.pa_id.isin(grpidx)]
df_OASIS = df_OASIS[df_OASIS.pa_id.isin(grpidx)]

df_OASIS.pa_id.nunique()

<hr style="height:2px;border-width:0;color:black;background-color:black">

<div class="alert alert-danger">
    <b>View datatypes - the int64 were causing a bug when serializing to JSON</b>
</div>

df_patients.dtypes#.head()
df_episodes.dtypes#.head()
df_visits.dtypes#.head()
df_OASIS.dtypes#.head()
df_event_mapping.dtypes#.head()

<hr style="height:2px;border-width:0;color:black;background-color:black">

# Model Setup and Data Processing

<div class="alert alert-warning">
    <b>Import packages needed to Parallel process the Data</b>
</div>

import joblib
from joblib import Parallel, delayed
import inspect

<div class="alert alert-warning">
    <b>View the number of cores recognized for parallel processing</b>
</div>


print(joblib.cpu_count() - 4)


<div class="alert alert-warning">
    <b>Import pyhealth classes and functions</b>
</div>

from pyhospice.data.base_mimic import parallel_parse_tables
from pyhospice.utils.utility_parallel import unfold_parallel
from pyhospice.utils.utility_parallel import partition_estimators
from pyhospice.utils.utility import read_csv_to_df
from pyhospice.utils.utility import make_dirs_if_not_exists

inspect.getsourcefile(parallel_parse_tables)
inspect.getsourcefile(unfold_parallel)
inspect.getsourcefile(partition_estimators)
inspect.getsourcefile(read_csv_to_df)
inspect.getsourcefile(make_dirs_if_not_exists)


<div class="alert alert-warning">
    <b>Create Output Directory to Save Files</b>
</div>

#from pathlib import Path
#Path("pyhealth_export").mkdir(parents=True, exist_ok=True)

<div class="alert alert-warning">
    <b>Setup the Parallel Proccesing Workflow</b>
</div>

save_dir = os.getcwd() + '/__exported'+ '/pyhealth_export'
n_jobz = (joblib.cpu_count() - 4)  # number of parallel jobs
duration = 21600  # time window for episode generation
selection_method = 'last'
var_list = df_event_mapping.OASIS.tolist()
patient_id_list = df_patients['pa_id'].tolist()

<div class="alert alert-danger">
    <b>Custom JSON Encoder to deal with np.int64 serialization Bug</b>
</div>

class NpEncoder(json.JSONEncoder):
    def default(self, obj):
        if isinstance(obj, np.integer):
            return int(obj)
        if isinstance(obj, np.floating):
            return float(obj)
        if isinstance(obj, np.ndarray):
            return obj.tolist()
        return super(NpEncoder, self).default(obj)

<div class="alert alert-warning">
    <b>Execute the Parallel Data Processing</b>
</div>

n_patients_list, starts, n_jobs = partition_estimators(
        len(patient_id_list), n_jobs=n_jobz)

all_results = Parallel(n_jobs=n_jobs, max_nbytes=None, verbose=True)(
delayed(parallel_parse_tables)(
     patient_df=df_patients,
     admission_df=df_episodes,
      icu_df=df_visits,
     event_df=df_OASIS,
     event_mapping_df=df_event_mapping,
     duration=duration,
     save_dir=save_dir,selection_method = selection_method, var_list = var_list, patient_id_list = patient_id_list)
 for i in range(n_jobs))

<div class="alert alert-warning">
    <b>Create JSON File With Patient-Level Data</b>
</div>

patient_data_loc = os.path.join(save_dir, 'patient_data_demo.json')
all_results = list(map(list, zip(*all_results)))
valid_data_list = unfold_parallel(all_results[0], n_jobs)
valid_id_list = unfold_parallel(all_results[1], n_jobs)

patient_data_list = []
for p in valid_data_list:
        patient_data_list.append(p.data)

with open(patient_data_loc, 'w') as outfile:
        json.dump(patient_data_list, outfile, cls=NpEncoder)

<div class="alert alert-warning">
    <b>Use Jupyter Magics which sends commands via Bash</b>
</div>

##### the data must be in these folders to work correctly

x_data = save_dir+'/x_data'
y_data = save_dir+'/y_data'
move = save_dir+'/*_episode.csv'

move

!mkdir $x_data
!mkdir $y_data

save_dir

!mv $save_dir*_episode.csv $x_data

#!Powershell.exe -Command "Move-Item -Path * -Filter *_episode.csv -Destination .\x_data"



<div class="alert alert-danger">
    <b>Examination of the Output</b>
</div>

type(all_results[0][0][0])
all_results[0][0][0]
patient_data_list

<div class="alert alert-warning">
    <b>Generate Labels for Mortality or Not</b>
</div>

os.path.exists(save_dir+'/y_data')

with open(patient_data_loc) as f:
    patient_data_list = json.load(f)
##########################################################
output_headers = ['episode_file', 'death_indicator']
output_df = pd.DataFrame(columns=output_headers)
ydir = save_dir+'/y_data'

for i, p_id in enumerate(tqdm(patient_data_list)):
    for adm in p_id['admission_list']:
        # csv file does not exist
        if 'episode_csv' in adm.keys():
            temp_list = [adm['episode_csv'], adm['death_indicator']]
            # append to the major episode dataframe
            temp_df = pd.DataFrame(temp_list).transpose()
            temp_df.columns = output_headers
            output_df = pd.concat([output_df, temp_df], axis=0)
# change file header to lower case

output_df.columns = output_df.columns.str.lower()
if not os.path.exists(ydir):
    os.makedirs(ydir)
output_df.to_csv(
        os.path.join(save_dir+'/y_data', 'mortality.csv'),
        index=False)

<hr style="height:2px;border-width:0;color:black;background-color:black">

<div class="alert alert-warning">
    <b>Import Python Package and Libraries needed for model execution</b>
</div>

from __future__ import division
from __future__ import print_function
from pathlib import Path
import torch # Needed if we will use GPU for model
print(torch.__version__)

<div class="alert alert-warning">
    <b>Import functions and classes from the pyhosice module</b>
</div>

from pyhospice.data.expdata_generator import sequencedata as expdata_generator
from pyhospice.evaluation.evaluator import func
from pyhospice.models.sequence.lstm import LSTM as model

<div class="alert alert-warning">
    <b>Verify the pyhosice model is importing from your Jupyter homebase</b>
</div>

inspect.getsourcefile(expdata_generator)
inspect.getsourcefile(func)
inspect.getsourcefile(model)

<div class="alert alert-warning">
    <b>Serialize the Data into Pickles for Model Input</b>
</div>

#data_dir = os.path.join(save_dir, 'model_output')

expdata_id = '04.16.mortality'
# set up the datasets
cur_dataset = expdata_generator(expdata_id, root_dir=save_dir)
cur_dataset.get_exp_data(sel_task='mortality', data_root=save_dir)
cur_dataset.load_exp_data()
cur_dataset.show_data()


<div class="alert alert-warning">
    <b>Run the Chosen Model!</b>
</div>

# initialize the model for training
# turn on GPU by setting use_gpu to True
expmodel_id = '04.16.mortality.LTSM.cpu'
clf = model(expmodel_id=expmodel_id, n_batchsize=64, use_gpu=False, n_epoch=10, gpu_ids='0, 1')
clf.fit(cur_dataset.train, cur_dataset.valid)

# load the best model for inference
clf.load_model()
clf.inference(cur_dataset.test)
pred_results = clf.get_results()
print(pred_results['hat_y'])

# evaluate the model
r = func(pred_results['hat_y'], pred_results['y'])

<div class="alert alert-warning">
    <b>View Results</b>
</div>

from sklearn.metrics import accuracy_score

accuracy_score(pred_results['y'], pred_results['hat_y'])

len(pred_results['y'])
len(pred_results['hat_y'])

pred_results['y'].tolist()

pred_results['hat_y'].tolist()

np.sum(np.abs(pred_results['y'] - pred_results['hat_y']))

r


np.sum(pred_results['hat_y'] * pred_results['y'])/ np.sum(pred_results['y']) if np.sum(pred_results['y'])>0 else 0.

np.sum(pred_results['hat_y'] * pred_results['y'])

np.sum(pred_results['y'])

pred_results['y'].tolist()

y_score = clf.decision_function(X_test)

 %load_ext autoreload

from sklearn.metrics import PrecisionRecallDisplay
display = PrecisionRecallDisplay.from_predictions(pred_results['y'], pred_results['hat_y'], name="LinearSVC")
_ = display.ax_.set_title("2-class Precision-Recall curve")

%matplotlib
from sklearn.metrics import precision_recall_curve
import matplotlib.pyplot as plt
#calculate precision and recall
precision, recall, thresholds = precision_recall_curve(pred_results['y'], pred_results['hat_y'])

#create precision recall curve
fig, ax = plt.subplots()
ax.plot(recall, precision, color='purple')

#add axis labels to plot
ax.set_title('Precision-Recall Curve')
ax.set_ylabel('Precision')
ax.set_xlabel('Recall')

#display plot
plt.show()

from sklearn.metrics import precision_recall_curve
from sklearn.metrics import PrecisionRecallDisplay

prec, recall, _ = precision_recall_curve(pred_results['y'], pred_results['hat_y'], pos_label=1)
pr_display = PrecisionRecallDisplay(precision=prec, recall=recall).plot()

import sklearn
sklearn.metrics.precision_recall_curve(pred_results['y'], pred_results['hat_y'], pos_label=1, sample_weight=None)

pred_results['hat_y'].tolist()

len(pred_results['hat_y'].tolist())

Compute Area Under the Receiver Operating Characteristic Curve (ROC AUC) from prediction scores.
The roc_auc_score function computes the area under the receiver operating characteristic (ROC) curve, which is also denoted by AUC
By computing the area under the roc curve, the curve information is summarized in one number.

clf.inference(cur_dataset.valid)
