==> 04_Debugging_Testing_Sandbox.ipynb <==
import pandas as  pd
import os
import numpy as np
import dtale
import json
from tqdm import tqdm
import sys

# %load C:\users\derek\.jupyter\startup.py
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

# from IPython.core.display import display, HTML
# display(HTML("<style>.container { width:100% !important; }</style>"))

#pd.set_option('display.max_colwidth', 100)

pd.set_option('display.max_colwidth', None)



pd.set_option("display.max_rows", None)

pd.options.display.max_columns = 500

pd.options.display.max_seq_items = 500



from pyhealth.data.base_mimic import * #parallel_parse_tables

df_patients = pd.read_csv('patients.csv')#.iloc[:1000,:]
df_episodes = pd.read_csv('df_episode.csv')#.iloc[:150,:]
df_visits = pd.read_csv('df_visits.csv')#.iloc[:150,:]
df_OASIS = pd.read_csv('df_OASIS.csv')#.iloc[:150,:]
df_event_mapping = pd.read_csv('event_mapping_df.csv')

var_list = df_event_mapping.OASIS.tolist()
save_dir = './testing2'
#duration = int(21600)

mimic = MIMIC_Data(df_patients.iloc[1][0], 21600, 'last')

mimic.parse_patient(df_patients)

mimic.parse_admission(df_episodes[df_episodes.pa_id == 16274])

mimic.parse_icu(df_visits[df_visits.pa_id == 16274])

a = mimic.parse_event(df_OASIS[df_OASIS.pa_id == 16274].reset_index(), save_dir, df_event_mapping, var_list)

a

  #max_time_diff = pd_df['secs_since_entry'].max()
    
int(np.ceil(60 / (21600/84200))

key_df = df_event_mapping[df_event_mapping['OASIS'].isin(var_list)]



key_df


tempdf = key_df.merge(a, left_on='itemID',right_on='itemID', suffixes=(None, '_delete')).drop(axis=1, columns=['index','OASIS_delete'], level=None, inplace=False, errors='raise')

tempdf

tempdf.sort_values(by='secs_since_entry', inplace=True)


for var in var_list:
                    print(var)
                    var_df = tempdf[tempdf['OASIS'] == var] # This is wrong
                    #var_df = temp_df[temp_df['variable'] == var]
                    # write records for each variable at interest
                   # temp_list = self.write_record(temp_list, var_df, var)


var_df

mimic.episode_headers_ = ['secs_since_entry']

mimic.episode_headers_

#  for var in var_list:
#             mimic.episode_headers_.extend([var, var + '_unit'])
#             #mimic.episode_headers_.append(var)
 # self.episode_headers_.extend([var, var + '_unit'])

mimic.generate_episode(a, mimic.time_duration, df_event_mapping, var_list)

a.secs_since_entry.nunique()[0]

a.secs_since_entry.unique()[1]

a.secs_since_entry.unique()

a.loc[a.secs_since_entry == a.secs_since_entry.unique()[1]]

mimic.data

df_event_mapping[df_event_mapping['OASIS'].isin(var_list)]

df_event_mapping.dtypes

var_list[0]

df_OASIS[df_OASIS.pa_id == 16274].reset_index()

df_patients.iloc[0].values[0]

df_episodes.where(df_episodes.pa_id == '16274').dropna()

for i, admission_event in enumerate(mimic.data['admission_list']): 
    print(i, admission_event)


df_OASIS['M0090_INFO_COMPLETED_DT'].max() 


(pd.to_datetime(df_OASIS['M0090_INFO_COMPLETED_DT']) - df[
                    'first_entry']).dt.total_seconds()
